Move load info into chip
